Finished voice channel functionality
Voice channels are created and destroyed when all users of the custom VC are gone.

+++ main.py
@@ -4,7 +4,7 @@ import os
 import json
 import random
 
-_id = 873267392233492481
+_id = 873565797887389696
 
 def findPrefix(client, message):
     with open ("prefixes.json", "r") as f:
@@ -31,10 +31,36 @@ async def on_voice_state_update(member, before, after):
     custom_channel_category = client.get_channel(873294766916386817)
     if not before.channel and after.channel:
         if after.channel.id == _id:
-            new_channel = await member.guild.create_voice_channel(name=f"{member.display_name}\'s Channel", category=custom_channel_category)
+            new_channel = await member.guild.create_voice_channel(
+                name=f"{member.display_name}\'s Channel",
+                category=custom_channel_category,
+                position = 0)
             await member.move_to(new_channel)
+            f = open("channels.json", "a+")
+            f.close()
+            with open ("channels.json", "r") as f:
+                try:
+                    channels = json.load(f)
+                except:
+                    channels = {}
+            with open ("channels.json", "w") as f:
+                channels[str(new_channel.id)] = member.id
+                json.dump(channels, f, indent=4)
     elif before.channel and not after.channel:
-        await before.channel.delete()
+        f = open("channels.json", "a+")
+        f.close()
+        with open ("channels.json", "r") as f:
+            channels = json.load(f)
+        with open ("channels.json", "w") as f:
+            try:
+                if str(before.channel.id) in channels:
+                    if before.channel.members == []:
+                        await before.channel.delete()
+                        channels.pop(str(before.channel.id))
+                json.dump(channels, f, indent=4)
+            except:
+                None
+
 
 @client.event
 async def on_ready():
